Removing the lock in Channel().
- The lock is used to protect appends/removes to the writequeue and readqueue
All single modifications to Python lists are atomic, because of the Global Interpreter Lock.

- The lock is used to protect self.ispoisoned in check_poison()
Since we don't have an unpoison action, then we should be safe without the lock

- The lock protects for loops in match() and poison()
But access to the lists can be made atomic, by creating a copy using writequeue[:] and readqueue[:]
And calls to the ChannelReq objects on the lists are protected themselves, so they are allowed to be interleaved.