Added buffered channels to all versions.
Use:  A = Channel(buffer=<N>) # where <N> is the wanted buffer size.

The Channel is implemented in buffer.py as a process maintaining a list of
buffered elements.

The alternative to implement buffered channels as buffered write requests
demanded handling many special cases, thus it polluted the readability of
the pycsp code.