Adding/confirming PyCSP windows support.
pycsp.greenlets is fully operational on both unix and windows platforms.
pycsp.parallel is fully operational on unix.

pycsp.parallel is operational on windows, but requires a few differences in usage if
the @multiprocess/MultiProcess process type is used:

 - Decorators is not allowed, because of the multiprocessing.Process implementation in windows.
   Instead, the classes Process, MultiProcess and Choice must be used directly.

 - The main body is executed multiple times, also because of the multiprocess.Process implementation.
   Thus, wrap the main code with the following check:
     if __name__ == '__main__':
       # main code


Applications written in a syntax compatible with CPython/windows will also be 
able to run in the CPython/unix environment.